Minor fixes (#68)
* Supporting df.write.bigquery implicit API
* fixing exception logging on missing path
* minor fixes

+++ src/main/scala/com/google/cloud/spark/bigquery/BigQueryWriteHelper.scala
@@ -19,9 +19,9 @@ import java.util.UUID
 
 import com.google.cloud.bigquery.{BigQuery, BigQueryException, JobInfo, LoadJobConfiguration}
 import com.google.cloud.http.BaseHttpServiceException
-import com.typesafe.scalalogging.StrictLogging
+import com.typesafe.scalalogging.Logger
 import org.apache.hadoop.conf.Configuration
-import org.apache.hadoop.fs.{Path, RemoteIterator}
+import org.apache.hadoop.fs.{FileSystem, Path, RemoteIterator}
 import org.apache.spark.sql.{DataFrame, SQLContext, SaveMode}
 
 import scala.collection.JavaConverters._
@@ -30,10 +30,9 @@ case class BigQueryWriteHelper(bigQuery: BigQuery,
                                sqlContext: SQLContext,
                                saveMode: SaveMode,
                                options: SparkBigQueryOptions,
-                               data: DataFrame) extends StrictLogging {
+                               data: DataFrame) {
 
-  import org.apache.spark.sql.SaveMode
-
+  private val logger = Logger(getClass)
   val conf = sqlContext.sparkContext.hadoopConfiguration
 
   val temporaryGcsPath = {
@@ -126,15 +125,27 @@ case class BigQueryWriteHelper(bigQuery: BigQuery,
  * @param conf the hadoop configuration
  */
 case class IntermediateDataCleaner(path: Path, conf: Configuration)
-  extends Thread
-    with StrictLogging {
+  extends Thread {
+  private val logger = Logger(getClass)
+
   def deletePath: Unit =
     try {
       val fs = path.getFileSystem(conf)
-      fs.delete(path, true)
+      if(pathExists(fs, path)) {
+        fs.delete(path, true)
+      }
     } catch {
       case e: Exception => logger.error(s"Failed to delete path $path", e)
     }
+
+  // fs.exists can throw exception on missing path
+  private def pathExists(fs: FileSystem, path: Path) : Boolean = {
+    try {
+      fs.exists(path)
+    } catch {
+      case e: Exception => false
+    }
+  }
 
   override def run : Unit = deletePath
 }
